REFACOT: 최근 검색어 삭제 API 코드 수정 및 검색 업데이트 추가

+++ src/main/java/com/there/src/search/SearchService.java
@@ -1,14 +1,13 @@
 package com.there.src.search;
 
-import com.there.config.BaseException;
-import com.there.config.BaseResponseStatus;
+import com.there.src.search.config.BaseException;
 import com.there.utils.JwtService;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import static com.there.config.BaseResponseStatus.DATABASE_ERROR;
+import static com.there.src.search.config.BaseResponseStatus.*;
 
 
 @Service
@@ -29,18 +28,19 @@ public class SearchService {
     }
 
 
-    // 최근 검색 삭제
-
-    public void deleteRecentSearch(int userIdx, int searchIdx) throws BaseException {
-
+    // 최근 검색어 삭제
+    public void deleteRecentSearch(int userIdx, int searchIdx) throws BaseException, com.there.config.BaseException {
 
         // 해당하는 유저의 검색 기록인지
+        if(searchProvider.checkUserSearchExist(userIdx, searchIdx) == 0){
+            throw new BaseException(USERS_SEARCHES_INVALID_ID);
+        }
 
         try{
 
             int result = searchDao.deleteRecentSearch(searchIdx);
             if(result == 0){
-//                trhow new BaseException(DELETE_FAIL_SEARCH);
+                throw new BaseException(DELETE_FAIL_SEARCH);
             }
 
         } catch (Exception exception) {
